New TraverserImpl
