--- src/internal/ds/index.ts
@@ -1,2 +1,0 @@
-export { PromiseQueue } from './PromiseQueue';
-export { SLLQueueExtended } from './SLLQueueExtended';
