+++ src/internal/implementations/index.ts
@@ -1,3 +1,3 @@
 export { BasicBatchMigratorImpl } from './BasicBatchMigratorImpl';
 export { BasicDefaultMigratorImpl } from './BasicDefaultMigratorImpl';
-export { PromiseQueueBasedTraverserImpl } from './PromiseQueueBasedTraverserImpl';
+export { TraverserImpl } from './TraverserImpl';
