+++ src/internal/utils/index.ts
@@ -1,4 +1,4 @@
-export { makeRetriable, registerInterval, sleep } from '@proficient/util';
+export { makeRetriable, sleep } from '@proficient/util';
 export { isTraverser } from './isTraverser';
 export * from './math';
 export * from './object';
